Refactor: moved patchApply to its own module common/orionldPatchApply.h/cpp
